Deleted obsolete LoginButton class. Moved Button class to LoginScreen package
