+++ src/LoginScreen/LoginPanel.java
@@ -21,7 +21,6 @@ import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.Timer;
 
-import Buttons.Button;
 import Stage.Commons;
 import Stage.ProjektFrame;
